# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClientProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  preferredLanguage?: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  timezone?: string | null;
}

export interface ClientProfileUpdate {
  name?: string;
  phone?: string;
  age?: number;
  gender?: string;
  preferredLanguage?: string;
}

export interface Therapist {
  id: number;
  name: string;
  title: string;
  /** @nullable */
  avatarUrl: string | null;
  yearsOfExperience?: number;
  specializations?: string[];
  languages?: string[];
  /** @nullable */
  bio?: string | null;
  isVerified: boolean;
  /** @nullable */
  rating?: number | null;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  upcoming: 'upcoming',
  past: 'past',
  cancelled: 'cancelled',
} as const;

export interface Session {
  id: number;
  status: SessionStatus;
  scheduledAt: string;
  durationMinutes: number;
  therapistName: string;
  /** @nullable */
  therapistAvatarUrl: string | null;
  /** @nullable */
  joinUrl?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ActivityCategory = typeof ActivityCategory[keyof typeof ActivityCategory];


export const ActivityCategory = {
  breathing: 'breathing',
  journaling: 'journaling',
  mindfulness: 'mindfulness',
  sleep: 'sleep',
  gratitude: 'gratitude',
  cbt: 'cbt',
  reading: 'reading',
} as const;

export type ActivityDifficulty = typeof ActivityDifficulty[keyof typeof ActivityDifficulty];


export const ActivityDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ActivityStatus = typeof ActivityStatus[keyof typeof ActivityStatus];


export const ActivityStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Activity {
  id: number;
  title: string;
  category: ActivityCategory;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  estimatedMinutes: number;
  completionPercent: number;
  difficulty: ActivityDifficulty;
  status: ActivityStatus;
  /** @nullable */
  reflection?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface ActivityReflection {
  reflection?: string;
}

export type AssessmentType = typeof AssessmentType[keyof typeof AssessmentType];


export const AssessmentType = {
  'PHQ-9': 'PHQ-9',
  'GAD-7': 'GAD-7',
  PANAS: 'PANAS',
  'PCL-5': 'PCL-5',
  AUDIT: 'AUDIT',
} as const;

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


export const AssessmentStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;

export interface ScoreHistoryPoint {
  date: string;
  score: number;
}

export interface Assessment {
  id: number;
  name: string;
  type: AssessmentType;
  /** @nullable */
  description?: string | null;
  status: AssessmentStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  estimatedMinutes: number;
  /** @nullable */
  score?: number | null;
  scoreHistory?: ScoreHistoryPoint[];
}

export interface AssessmentSubmission {
  score: number;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface GoalProgress {
  name: string;
  current: number;
  target: number;
}

export interface ProgressData {
  activitiesCompleted: number;
  currentStreak: number;
  goalsAchieved: number;
  attendanceRate: number;
  wellnessScore: number;
  moodTrend: TrendPoint[];
  activityCompletion: TrendPoint[];
  assessmentTrends: TrendPoint[];
  goalProgress?: GoalProgress[];
}

export type ResourceCategory = typeof ResourceCategory[keyof typeof ResourceCategory];


export const ResourceCategory = {
  article: 'article',
  video: 'video',
  worksheet: 'worksheet',
  meditation: 'meditation',
  pdf: 'pdf',
} as const;

export interface Resource {
  id: number;
  title: string;
  category: ResourceCategory;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  readingMinutes?: number | null;
  /** @nullable */
  author?: string | null;
  isSaved: boolean;
  isSharedByTherapist?: boolean;
  /** @nullable */
  downloadUrl?: string | null;
}

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  therapist: 'therapist',
  reminder: 'reminder',
  appointment: 'appointment',
  notification: 'notification',
} as const;

export interface Message {
  id: number;
  type: MessageType;
  /** @nullable */
  senderId: number | null;
  senderName: string;
  /** @nullable */
  senderAvatarUrl?: string | null;
  content: string;
  sentAt: string;
  isRead: boolean;
}

export interface MessageInput {
  content: string;
}

export interface DashboardSummary {
  clientName: string;
  activitiesCompleted: number;
  currentStreak: number;
  goalsAchieved: number;
  upcomingSession: Session;
  todayTasks: Activity[];
  recentMessage: Message;
  sharedResources: Resource[];
}

export type GetSessionsParams = {
status?: GetSessionsStatus;
};

export type GetSessionsStatus = typeof GetSessionsStatus[keyof typeof GetSessionsStatus];


export const GetSessionsStatus = {
  upcoming: 'upcoming',
  past: 'past',
  cancelled: 'cancelled',
} as const;

export type GetResourcesParams = {
category?: GetResourcesCategory;
};

export type GetResourcesCategory = typeof GetResourcesCategory[keyof typeof GetResourcesCategory];


export const GetResourcesCategory = {
  article: 'article',
  video: 'video',
  worksheet: 'worksheet',
  meditation: 'meditation',
  pdf: 'pdf',
} as const;