# lib/api-zod/src/generated/types/clientProfileUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientProfileUpdate {
  name?: string;
  phone?: string;
  age?: number;
  gender?: string;
  preferredLanguage?: string;
}