# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Message } from './message';
import type { Resource } from './resource';
import type { Session } from './session';

export interface DashboardSummary {
  clientName: string;
  activitiesCompleted: number;
  currentStreak: number;
  goalsAchieved: number;
  upcomingSession: Session;
  todayTasks: Activity[];
  recentMessage: Message;
  sharedResources: Resource[];
}