# lib/api-zod/src/generated/types/goalProgress.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalProgress {
  name: string;
  current: number;
  target: number;
}