# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';

export interface Resource {
  id: number;
  title: string;
  category: ResourceCategory;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  readingMinutes?: number | null;
  /** @nullable */
  author?: string | null;
  isSaved: boolean;
  isSharedByTherapist?: boolean;
  /** @nullable */
  downloadUrl?: string | null;
}