# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hexpertify Client Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  label: string;
  value: number;
}